+++ tests/Sniffs/CompositeCodeElementSniffTest.php
@@ -159,7 +159,12 @@ class CompositeCodeElementSniffTest extends TestCase
                 '092 Use void return value type declaration or change type to compound, e.g. SomeClass|null',
                 '094 Add type declaration for return value, e.g.: "?SomeClass"',
                 '101 Useless PHPDoc',
-                '117 Replace array type with typed array type in PHPDoc for C4 constant, .e.g.: "string[]" or "SomeClass[]". Use mixed[] for generic arrays. Correct array depth must be specified.'
+                '107 Use a more specific type in typed array hint "[]" for parameter $arg1. Correct array depth must be specified.',
+                '118 Replace array type with typed array type in PHPDoc for C4 constant, .e.g.: "string[]" or "SomeClass[]". Use mixed[] for generic arrays. Correct array depth must be specified.',
+                '121 Use a more specific type in typed array hint "[]" for C5 constant. Correct array depth must be specified.',
+                '124 Use a more specific type in typed array hint "[][]" for C6 constant. Correct array depth must be specified.',
+                '124 Type hint "null" is not compatible with C6 constant value type',
+                '127 Use a more specific type in typed array hint "[][]" for property $prop2. Correct array depth must be specified.',
             ],
         ];